requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyautogui==0.9.54
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
psutil==5.9.5
pywin32==306
mss==9.0.1
uuid==1.30
requests==2.31.0 